+++ src/edu/csupomona/cs/cs356/classmate/fragments/friends/FriendListAdapter.java
@@ -121,7 +121,7 @@ public class FriendListAdapter extends ArrayAdapter<Friend> implements View.OnCl
                 
                 if (getContext() instanceof FragmentActivity) {
                     // We can get the fragment manager
-                        int id = f.getID();
+                        long id = f.getID();
                         Fragment newFragment = new edu.csupomona.cs.cs356.classmate.fragments.ScheduleFragment(id);
                     FragmentActivity activity = ((FragmentActivity)getContext());
                     FragmentTransaction t = activity.getSupportFragmentManager().beginTransaction();
@@ -148,7 +148,7 @@ public class FriendListAdapter extends ArrayAdapter<Friend> implements View.OnCl
 
                                 RequestParams params = new RequestParams();
                                 params.put("email", emailAddress);
-                                params.put("user_id", Integer.toString(f.getID()));
+                                params.put("user_id", Long.toString(f.getID()));
                                 params.put("version", "1");
 
                                 AsyncHttpClient client = new AsyncHttpClient();
@@ -169,7 +169,7 @@ public class FriendListAdapter extends ArrayAdapter<Friend> implements View.OnCl
 
                                 RequestParams params = new RequestParams();
                                 params.put("email", emailAddress);
-                                params.put("user_id", Integer.toString(f.getID()));
+                                params.put("user_id", Long.toString(f.getID()));
                                 params.put("version", "2");
 
                                 AsyncHttpClient client = new AsyncHttpClient();
