+++ src/edu/csupomona/cs/cs356/classmate/fragments/friends/FriendRequestsTab.java
@@ -46,7 +46,7 @@ public class FriendRequestsTab extends Fragment {
 						JSONArray myjsonarray = new JSONArray(response);
 						for (int i = 0; i < myjsonarray.length(); i++) {
 							jObj = myjsonarray.getJSONObject(i);
-							requests.add(new Friend(jObj.getInt("user_id"), jObj.getString("username"), jObj.getString("email")));
+							requests.add(new Friend(jObj.getLong("user_id"), jObj.getString("username"), jObj.getString("email")));
 						}
 					} catch (JSONException e) {
 						e.printStackTrace();
