+++ src/edu/csupomona/cs/cs356/classmate/fragments/groups/AddGroupTab.java
@@ -47,11 +47,11 @@ public class AddGroupTab extends Fragment {
 				llProgressBar.setVisibility(View.VISIBLE);
 				lvSearchResults.setAdapter(null);
 
-				int id = getActivity().getIntent().getIntExtra(INTENT_KEY_USERID, NULL_USER);
+				long id = getActivity().getIntent().getLongExtra(INTENT_KEY_USERID, NULL_USER);
 
 				RequestParams params = new RequestParams();
 				params.put("search", e.toString());
-				params.put("user_id", Integer.toString(id));
+				params.put("user_id", Long.toString(id));
 
 				AsyncHttpClient client = new AsyncHttpClient();
 				client.get("http://www.lol-fc.com/classmate/searchgroups.php", params, new AsyncHttpResponseHandler() {
