+++ src/edu/csupomona/cs/cs356/classmate/fragments/groups/AddMemberActivity.java
@@ -63,7 +63,7 @@ public class AddMemberActivity extends Activity {
 								for (int i = 0; i < myjsonarray.length(); i++) {
 									jObj = myjsonarray.getJSONObject(i);
 									search_results.add(new Friend(
-										jObj.getInt("user_id"),
+										jObj.getLong("user_id"),
 										jObj.getString("username"),
 										jObj.getString("email")
 									));
